project 및 milestone 충돌 해결

+++ src/main/java/com/nhnacademy/taskApi/dto/project/request/ProjectGetResponse.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 public class ProjectGetResponse {
     private Long projectId;
     private String projectName;
-    private Long statusId;
+    private Integer statusId;
 }
